Changes as per comments
