+++ app/res/navigation/nav_graph_connect.xml
@@ -108,7 +108,7 @@
             app:popUpTo="@id/connect_jobs_list_fragment" />
         <action
             android:id="@+id/action_connect_job_delivery_progress_fragment_to_connect_delivery_fragment"
-            app:destination="@id/connect_delivery_visits_fragment" />
+            app:destination="@id/connect_delivery_visits_detail_fragment" />
         <action
             android:id="@+id/action_connect_job_delivery_progress_fragment_to_connect_job_detail_bottom_sheet_dialog_fragment"
             app:destination="@id/connect_job_detail_bottom_sheet_dialog_fragment" />
@@ -124,8 +124,8 @@
             app:destination="@id/connect_downloading_fragment"
             app:popUpTo="@id/connect_jobs_list_fragment" />
         <action
-            android:id="@+id/action_connect_delivery_opp_home_fragment_to_connect_delivery_visits_fragment"
-            app:destination="@id/connect_delivery_visits_fragment" />
+            android:id="@+id/action_connect_delivery_opp_home_fragment_to_connect_delivery_visits_detail_fragment"
+            app:destination="@id/connect_delivery_visits_detail_fragment" />
         <action
             android:id="@+id/action_connect_delivery_opp_home_fragment_to_connect_job_detail_bottom_sheet_dialog_fragment"
             app:destination="@id/connect_job_detail_bottom_sheet_dialog_fragment" />
@@ -153,10 +153,10 @@
     </fragment>
 
     <fragment
-        android:id="@+id/connect_delivery_visits_fragment"
-        android:name="org.commcare.fragments.connect.ConnectDeliveryVisitsFragment"
-        android:label="fragment_connect_delivery_visits"
-        tools:layout="@layout/fragment_connect_delivery_visits">
+        android:id="@+id/connect_delivery_visits_detail_fragment"
+        android:name="org.commcare.fragments.connect.ConnectDeliveryVisitsDetailFragment"
+        android:label="fragment_connect_delivery_visits_detail"
+        tools:layout="@layout/fragment_connect_delivery_visits_detail">
         <argument
             android:name="unitId"
             app:argType="string" />
