+++ app/src/org/commcare/fragments/RefreshableTab.kt
@@ -1,0 +1,5 @@
+package org.commcare.fragments
+
+interface RefreshableTab {
+    fun updateView()
+}
